adding some first implementations
