+++ src/memory/paging/mod.rs
@@ -1,8 +1,9 @@
 //! Includes the different paging implementation.
+mod alloc;
 mod frame_allocator;
 mod physical_mmap;
 mod utils;
-mod alloc;
+mod virtual_mmap;
 
 use core::{arch::asm, marker::PhantomData, ops::Range};
 
@@ -15,7 +16,7 @@ use x86_64::{
     PhysAddr, VirtAddr,
 };
 
-use self::{frame_allocator::FRAME_ALLOCATOR, utils::table_wrapper::TableWrapper};
+use self::{frame_allocator::FRAME_ALLOCATOR, utils::table_wrapper::TableWrapper, virtual_mmap::{VMMMapper, VMMAPPER}};
 
 use crate::memory::{paging::physical_mmap::KernelData, HHDM};
 
@@ -26,7 +27,7 @@ lazy_static::lazy_static! {
     pub static ref HEAP_START: VirtAddr = VirtAddr::new(0x1000);
 }
 
-pub static PML4E_ADDR: Once<VirtAddr> = Once::new();
+pub static PML4E_ADDR: Once<PhysAddr> = Once::new();
 
 /// The amount of pages which should be used in the beginning for the stack.
 /// == 64KiB
@@ -34,6 +35,14 @@ const STACK_INIT_PAGES: u64 = 16;
 pub static STACK_START: Once<VirtAddr> = Once::new();
 
 pub fn init() -> ! {
+    let pml4e_addr = FRAME_ALLOCATOR
+        .write()
+        .allocate_frame()
+        .unwrap()
+        .start_address();
+
+    PML4E_ADDR.call_once(|| pml4e_addr.clone());
+
     let p_configurator = KPagingConfigurator::<Size4KiB>::new();
     p_configurator.map_kernel();
     p_configurator.map_heap();
@@ -66,19 +75,12 @@ pub struct KPagingConfigurator<P: PageSize> {
 impl<P: PageSize> KPagingConfigurator<P> {
     /// Creates a new pornos-paging-configurator
     pub fn new() -> Self {
-        let pml4e_addr = FRAME_ALLOCATOR
-            .write()
-            .allocate_frame()
-            .unwrap()
-            .start_address();
-
-        PML4E_ADDR.call_once(|| *HHDM + pml4e_addr.as_u64());
-        let ptr = PML4E_ADDR.get().unwrap().as_mut_ptr() as *mut PageTable;
+        let p4_phys_addr = PML4E_ADDR.get().unwrap().clone();
 
         Self {
             size: PhantomData,
-            p4_phys_addr: pml4e_addr,
-            p4_ptr: ptr,
+            p4_phys_addr,
+            p4_ptr: p4_phys_addr.as_u64() as * mut PageTable,
         }
     }
 
@@ -179,76 +181,6 @@ impl<P: PageSize> KPagingConfigurator<P> {
 }
 
 impl<P: PageSize> KPagingConfigurator<P> {
-    /// Maps a page to the given page_frame (if available) with the given flags.
-    ///
-    /// * `page`: The page to be mapped.
-    /// * `page_frame`: If it's `Some`, then the page will be mapped to the given page frame,
-    ///                 otherwise a new page frame will ba allocated.
-    /// * `flags`: The flags for the given mapping.
-    pub fn map_page(&self, page: Page, page_frame: Option<PhysFrame>, flags: PageTableFlags) {
-        let mut table_wrapper = TableWrapper::new(self.p4_ptr);
-        let mut level = PageTableLevel::Four;
-
-        while let Some(lower_level) = level.next_lower_level() {
-            let entry_index = match lower_level {
-                PageTableLevel::Three => page.start_address().p4_index(),
-                PageTableLevel::Two => page.start_address().p3_index(),
-                PageTableLevel::One => page.start_address().p2_index(),
-                _ => unreachable!("Ayo, '{:?}' shouldn't be here <.<", lower_level),
-            };
-            let table_entry = table_wrapper.get_entry(entry_index);
-
-            let next_table_ptr = {
-                let next_table_vtr_ptr = if table_entry.is_unused() {
-                    let flags = PageTableFlags::WRITABLE | PageTableFlags::PRESENT;
-                    table_wrapper.set_page_frame(entry_index, None, flags);
-                    *HHDM + table_wrapper.get_entry(entry_index).addr().as_u64()
-                } else {
-                    *HHDM + table_entry.addr().as_u64()
-                };
-                next_table_vtr_ptr.as_mut_ptr() as *mut PageTable
-            };
-
-            table_wrapper = TableWrapper::new(next_table_ptr);
-            level = lower_level;
-        }
-
-        table_wrapper.set_page_frame(page.p1_index(), page_frame, flags);
-    }
-
-    /// Maps a range of pages in a romw.
-    ///
-    /// * `page`: The starting page which should be mapped.
-    /// * `page_frame`: The starting page frame (if available) which should be mapped.
-    ///                 If it's `None`, random page-frames are picked up then.
-    /// * `len`: The amount of bytes which should be mapped in a row.
-    /// * `flags`: The flags for each page.
-    ///
-    /// # Note
-    /// If `page_frame` is `Some(...)`, then you **have to** make sure that, the range, starting
-    /// from the given page frame until `start + len` is **a valid Physicall address range**!!!
-    pub fn map_page_range(
-        &self,
-        page: Page,
-        page_frame: Option<PhysFrame>,
-        len: Bytes,
-        flags: PageTableFlags,
-    ) {
-        for offset in (0..len.as_u64()).step_by(P::SIZE.try_into().unwrap()) {
-            let page = {
-                let addr = (page.start_address() + offset).align_down(P::SIZE);
-                Page::from_start_address(addr).unwrap()
-            };
-
-            let page_frame = page_frame.map(|frame| {
-                let addr = (frame.start_address() + offset).align_down(P::SIZE);
-                PhysFrame::from_start_address(addr).unwrap()
-            });
-
-            self.map_page(page, page_frame, flags);
-        }
-    }
-
     pub fn map_kernel_part(
         &self,
         phys_kernel_start: PhysAddr,
